requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
pandas
numpy
sentence-transformers
qdrant-client
pytest
pytest-cov